separated the models into separate classes and added an EA model
